requirements.txt
langchain 
ipykernel # for selecting kernel fo jupyter notebook file 
python-dotenv 
langchain_community 
pypdf 
bs4
arxiv
pymupdf 
wikipedia
langchain-text-splitters 
lxml
langchain-openai
chromadb
langchain_community

sentence_transformers 



faiss-cpu 

langchain-chroma

streamlit
fastapi

langserve

sse_starlette 
pydantic

langchain_huggingface
huggingface_hub